+++ src/priorityqueue/PriorityQueue.java
@@ -24,7 +24,7 @@ public interface PriorityQueue<T> {
      * @param priority
      * @throws QueueOverflowException
      */
-    public void add(T item, int priority) throws QueueOverflowException;
+    public void add(T item, long priority) throws QueueOverflowException;
 
     /**
      * The highest priority item stored. Throw an exception if it's empty.
